Split event builders into separate files
