+++ src/platforms/alexa/apis/proactiveEventsBuilders/OrderStatusEventBuilder.ts
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2018 Rain Agency <contact@rain.agency>
+ * Author: Rain Agency <contact@rain.agency>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import { EventsBuilder } from "../ProactiveEvents";
+
+/**
+ * Order Status Events Builder class reference
+ */
+export class OrderStatusEventBuilder extends EventsBuilder {
+  public state: any = {};
+
+  public setStatus(status: ORDER_STATUS, expectedArrival?: string, enterTimestamp?: string): OrderStatusEventBuilder {
+    this.state = { status, enterTimestamp };
+
+    if (expectedArrival) {
+      this.state.deliveryDetails = { expectedArrival };
+    }
+
+    return this;
+  }
+
+  public build(): OrderStatusEventBuilder {
+    const payload = {
+      order: {
+        seller: {
+          name: "localizedattribute:sellerName",
+        },
+      },
+      state: this.state,
+    };
+
+    this.setName("AMAZON.OrderStatus.Updated");
+    this.setPayload(payload);
+
+    return super.build();
+  }
+}
+
+export enum ORDER_STATUS {
+  ORDER_DELIVERED = "ORDER_DELIVERED",
+  ORDER_OUT_FOR_DELIVERY = "ORDER_OUT_FOR_DELIVERY",
+  ORDER_PREPARING = "ORDER_PREPARING",
+  ORDER_RECEIVED = "ORDER_RECEIVED",
+  ORDER_SHIPPED = "ORDER_SHIPPED",
+  PREORDER_RECEIVED = "PREORDER_RECEIVED",
+}
